Final_Working_Model/requirements.txt
Flask>=2.0
GitPython
joblib
pandas
scikit-learn
